+++ app/models/early_career_payments/eligibility_answers_presenter.rb
@@ -19,11 +19,20 @@ module EarlyCareerPayments
     # [2]: slug for changing the answer.
     def answers
       [].tap do |a|
+        a << has_entire_term_contract
         a << nqt_in_academic_year_after_itt
       end
     end
 
     private
+
+    def has_entire_term_contract
+      [
+        translate("early_career_payments.questions.has_entire_term_contract"),
+        (eligibility.has_entire_term_contract? ? "Yes" : "No"),
+        "entire-term-contract"
+      ]
+    end
 
     def nqt_in_academic_year_after_itt
       [