+++ spec/models/claim_spec.rb
@@ -905,7 +905,7 @@ RSpec.describe Claim, type: :model do
   end
 
   describe "Early Career Payments claim" do
-    let(:eligibility) { build(:early_career_payments_eligibility) }
+    let(:eligibility) { build(:early_career_payments_eligibility, :eligible) }
 
     describe "#submittable?" do
       it "returns true when the claim is valid and has not been submitted" do
@@ -913,6 +913,7 @@ RSpec.describe Claim, type: :model do
 
         expect(claim.submittable?).to eq true
       end
+
       it "returns false when it has already been submitted" do
         claim = build(:claim, :unverified, eligibility: eligibility)
 
@@ -921,7 +922,7 @@ RSpec.describe Claim, type: :model do
     end
 
     it "triggers validations on the eligibility appropriate to the context" do
-      claim = build(:claim, eligibility: eligibility)
+      claim = build(:claim, eligibility: build(:early_career_payments_eligibility))
 
       expect(claim).not_to be_valid(:"nqt-in-academic-year-after-itt")
       expect(claim.errors.values).to include(["Select yes if you did your NQT in the academic year after your ITT"])